requirements.txt
agno
python-dotenv
duckduckgo-search
lancedb
pandas
groq
uvicorn
fastapi
openai 